feat(architect): geometry methods
Geometry methods can be called when creating some kind of shape.
We should pass the VBO and VAO (optionally the EBO) back to Ruby, so we
can then use those to draw the geometry each frame.

VBOs, VAOs and EBOs are not affected by the Ruby garbage collector (the
free method is NULL). We need to manually clean them up.